+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.09] - 2024-11-28
+
+### Added
+
+- Add correct full login process
+- Add complete login UI and routes
+- Add remove and generate recovery tokens
+- Add forgot password fix
+
 ## [0.2.08] - 2024-11-26
 
 ### Added